volk_rank_archs: widen impl_deps to uint64_t, fix name matching
The architecture dispatcher used int (32-bit signed) for implementation
dependency bitmasks, which overflows when ISA flag indices exceed 30.
With 28 architectures currently defined, only 3 spare bits remain.
Implementation name matching truncated at 20 characters via strncmp,
and volk_get_index recursed infinitely when an implementation was not
found.

Widen impl_deps from int to uint64_t in volk_rank_archs and all
codegen templates. Widen volk_get_lvarch return type and volk_machine
caps field to uint64_t. Use UINT64_C(1) for all ISA flag shifts.
Replace strncmp with strcmp for exact name matching. Replace the
recursive fallback in volk_get_index with a linear search for
"generic", then index 0.

Fixes https://github.com/mtibbits/volk/issues/16
Related: https://github.com/gnuradio/volk/issues/516

+++ tmpl/volk.tmpl.c
@@ -29,7 +29,7 @@ struct volk_machine *get_machine(void)
   if(machine != NULL)
     return machine;
   else {
-    unsigned int max_score = 0;
+    uint64_t max_score = 0;
     unsigned int i;
     struct volk_machine *max_machine = NULL;
     for(i=0; i<n_volk_machines; i++) {
@@ -71,7 +71,7 @@ const char* volk_get_machine(void)
   if(machine != NULL)
     return machine->name;
   else {
-    unsigned int max_score = 0;
+    uint64_t max_score = 0;
     unsigned int i;
     struct volk_machine *max_machine = NULL;
     for(i=0; i<n_volk_machines; i++) {
@@ -133,7 +133,7 @@ static inline void __init_${kern.name}(void)
 {
     const char *name = get_machine()->${kern.name}_name;
     const char **impl_names = get_machine()->${kern.name}_impl_names;
-    const int *impl_deps = get_machine()->${kern.name}_impl_deps;
+    const uint64_t *impl_deps = get_machine()->${kern.name}_impl_deps;
     const bool *alignment = get_machine()->${kern.name}_impl_alignment;
     const size_t n_impls = get_machine()->${kern.name}_n_impls;
     const size_t index_a = volk_rank_archs(name, impl_names, impl_deps, alignment, n_impls, true/*aligned*/);
@@ -188,7 +188,7 @@ void ${kern.name}_manual(${kern.arglist_full}, const char* impl_name)
 
 volk_func_desc_t ${kern.name}_get_func_desc(void) {
     const char **impl_names = get_machine()->${kern.name}_impl_names;
-    const int *impl_deps = get_machine()->${kern.name}_impl_deps;
+    const uint64_t *impl_deps = get_machine()->${kern.name}_impl_deps;
     const bool *alignment = get_machine()->${kern.name}_impl_alignment;
     const size_t n_impls = get_machine()->${kern.name}_n_impls;
     volk_func_desc_t desc = {
